Trying to get the two tab bar views to not go underneath the navigation bar.
Still need to experiment with it. I might need to add navigation bars
to each of the two view controllers that will appear beneath (and be
hidden by) the navigation bar on the MapAndTable tab bar controller.
Or, I could get rid of that one and create bars on the other two.