Fixed the Refresh button crash.
The flag on the tab bar used to check whether a new location was posted
was an optional. If you hadn’t gone to the info view, there was no way
to set that flag, so if you clicked on the Refresh button, the app
would crash unwrapping the flag. All I had to do was give it an initial
value and everything else still worked fine.