The login view controller looks better on the iPhone 5s simulator.
I also hid the Facebook button, as per the suggestion from the code
reviewer. I didn’t implement Facebook functionality, since I don’t have
an account (that’s what the alert said when you clicked on the button),
and the recommendation was just to get rid of it, or hide it.